Steering wheel and motor
I created a systsem using polymorphism where different types of scripts affect different aspects of the boat. Steering wheel affects, well, steering and the motor affects speed, acceleration, drag, etc.

It is made in a way so that i can create multiple different variations of boat components and the same structure should be used for when making other components such as the hull or lights.

It is important to note though that this was created for testing via computer and isnt adapted to vr yet. Also the boat moves by altering its world position, meaning that eventuall collisions with other objects (INCLUDING THE PLAYER) will have to be worked around to achive.